requirements.txt
fastapi
uvicorn
binance-connector
pandas
python-dotenv
httpx
apscheduler
